Added version 2.0 and class file
Update containing the second version of the script, in which I shuffled things around, moved most functions in a Thread class (in object.py), and added docstrings with descriptions of what everything does. Hopefully the code is now cleaner and easier to read, and also easier to update and debug.